requirements.txt
openai>=1.0.0
azure-identity>=1.15.0
pyautogui>=0.9.54
Pillow>=10.0.0
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
python-dotenv>=1.0.0
requests>=2.31.0
